backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.0
google-generativeai>=0.8.0
anthropic>=0.40.0
